% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{as_rect}
\alias{as_rect}
\alias{as_rect.grout_tiles}
\title{Create rect (extent) table}
\usage{
as_rect(x, ...)

\method{as_rect}{grout_tiles}(x, ...)
}
\arguments{
\item{x}{a grout [tiles()] object}

\item{...}{ignored}
}
\value{
data frame of tile extents
}
\description{
Data frame of extents xmin, xmax, ymin, ymax.
}
