% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{plot.grout_tiles}
\alias{plot.grout_tiles}
\title{Plot a tiling scheme}
\usage{
\method{plot}{grout_tiles}(x, ..., add = FALSE, border = "grey", lwd = 2)
}
\arguments{
\item{x}{a `"grout_tiles"` object from [grout()].}

\item{...}{passed to [vaster::plot_extent()].}

\item{add}{add to the current plot?  Default `FALSE`.}

\item{border}{colour for the tile borders.  Default `"grey"`.}

\item{lwd}{line width for tile borders.  Default `2`.}
}
\description{
Draws each tile as a rectangle (grey border) with the original raster
extent overlaid as a dashed red outline.
}
