% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropo.R
\name{plot.tropo_tiles}
\alias{plot.tropo_tiles}
\title{plot tiles}
\usage{
\method{plot}{tropo_tiles}(x, ..., border = "grey", lwd = 2)
}
\arguments{
\item{x}{a tropo [tiles()] object}

\item{...}{arguments passed to methods (particularly [sp::plot()])}

\item{border}{the colour of the tile border (default grey)}

\item{lwd}{the width of the tile border (default = 2)}
}
\description{
plot tiles
}
