% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{print.grout_tiles}
\alias{print.grout_tiles}
\title{print tiles}
\usage{
\method{print}{grout_tiles}(x, ...)
}
\arguments{
\item{x}{a grout [tiles()] object}

\item{...}{ignored}
}
\description{
print tiles
}
