% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropo.R
\name{print.tropo_tiles}
\alias{print.tropo_tiles}
\title{print tiles}
\usage{
\method{print}{tropo_tiles}(x, ...)
}
\arguments{
\item{x}{a tropo [tiles()] object}

\item{...}{ignored}
}
\description{
print tiles
}
