% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropo.r
\name{tiles}
\alias{tiles}
\title{tiles of a raster}
\usage{
tiles(x, blockX = 256, blockY = 256)
}
\arguments{
\item{x}{a raster}

\item{blockX}{tile dimensions in columns (X)}

\item{blockY}{tile diemnsions in rows (Y)}
}
\value{
A raster layer with cells having the tile number and extent releant to the input.
}
\description{
tiles of a raster
}

