% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropo.R
\name{as_polys}
\alias{as_polys}
\alias{as_polys.grout_tiles}
\title{Tiles as polygons}
\usage{
as_polys(x, ...)

\method{as_polys}{grout_tiles}(x, ...)
}
\arguments{
\item{x}{a grout [tiles()] object}

\item{...}{ignored}
}
\description{
Tiles as polygons
}
