% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grout-package.R
\docType{package}
\name{grout-package}
\alias{grout-package}
\title{grout: Abstract Raster Tiling Schemes}
\description{
Impose a tiling scheme on a raster grid defined by its dimension and extent. Computes tile counts, pixel offsets (suitable for GDAL ReadRaster calls), and geographic extents for each tile. Also supports standard Web Mercator and geodetic tiling profiles with zoom levels, and provides a spatial map of what each chunk covers for interrogating virtual or virtualizable raster stores.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/grout}
  \item Report bugs at \url{https://github.com/hypertidy/grout/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})

Authors:
\itemize{
  \item Michael D. Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})
}

}
\keyword{internal}
