% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grout-package.R
\docType{package}
\name{grout-package}
\alias{grout}
\alias{grout-package}
\title{grout: Create Raster Tiles}
\description{
Create raster tiles abstractly, or from large input files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/grout}
  \item Report bugs at \url{https://github.com/hypertidy/grout/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
