% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\name{plot.grout_tiles}
\alias{plot.grout_tiles}
\title{plot tiles}
\usage{
\method{plot}{grout_tiles}(x, ..., border = "grey", lwd = 2)
}
\arguments{
\item{x}{a grout [tiles()] object}

\item{...}{arguments passed to methods}

\item{border}{the colour of the tile border (default grey)}

\item{lwd}{the width of the tile border (default = 2)}
}
\description{
plot tiles
}
