% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offsets.R
\name{tile_index}
\alias{tile_index}
\title{Tile index}
\usage{
tile_index(x)
}
\arguments{
\item{x}{tiling object, created by [tiles()]}
}
\value{
data frame, see details
}
\description{
Generate a table of tile indexes from a tiling object.
}
\details{
A data frame with 

* **tile** - index, same as raster cell number
* **offset_x** - column offset of tile (index column 0-based)
* **offset_y** - row offset of tile (index row 0-based, relative to top row)
* **ncol** number of columns in tile (the right and bottom margins may have a dangle based on block size)
* **nrow** number of rows in th tile 
* xmin, xmax, ymin, ymax - the exent

Note that ncol,nrow is the block size *unless* the tile is part of a dangling column or row (right or bottom) where
the raster doesn't fill complete tiles.
}
\examples{
tile_index(grout(c(87, 61), extent = c(0, 1, 0, 1), blocksize = c(32, 16)))
## only one tile in this weird scheme!
tile_index(grout(c(61, 87), blocksize = c(61, 87)))
}
