% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_spec.R
\name{tile_spec}
\alias{tile_spec}
\alias{tile_zoom}
\title{Create tile specification}
\usage{
tile_spec(
  dimension,
  extent,
  zoom = 0,
  blocksize = c(256L, 256L),
  profile = c("mercator", "geodetic", "raster"),
  crs = NA,
  xyz = FALSE
)

tile_zoom(
  dimension,
  extent,
  blocksize = c(256L, 256L),
  profile = c("mercator", "geodetic", "raster")
)
}
\arguments{
\item{dimension}{size in pixels ncol,nrow}

\item{extent}{xmin,xmax,ymin,ymax}

\item{zoom}{the zoom level, starts at 0 and can be up to 24}

\item{blocksize}{size of each tile, defaults to 256x256}

\item{profile}{profile domain to use, see Details}

\item{crs}{crs, for raster profile}

\item{xyz}{default is `FALSE`, if `TRUE` use xyz-mode (zero is at the top)}
}
\value{
data frame with tile specification, tile index, tile_col, tile_row, 
 ncol, nrow, xmin, xmax, ymin, ymax, crs
}
\description{
Take an input grid (dimension and extent) and create a specification of
the tiling for that grid within a profile.
}
\details{
Profiles are 'mercator' or "geodetic" for global systems, or can 
use "raster" which will use the input grid specification as the entire domain. 

'tile_row' in output is in TMS orientation (zero is at the bottom), use 
'xyz' arg to switch. 

Function `tile_zoom()` will return the "natural" maximum zoom level, i.e. 
the zoom at which the tile resolution is just below the input resolution. 
Note that no reprojection is done, the input extent must match the profile chosen (use 'raster' for native profile).
}
\examples{
tile_spec(c(8194, 8194), c(140, 155, -45, -30), profile = "geodetic")

tile_spec(c(2048, 248), c(140, 155, -45, -30), zoom = 5, profile = "geodetic", 
   blocksize = c(512, 512))
}
