% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropo-package.R
\docType{package}
\name{tropo-package}
\alias{tropo}
\alias{tropo-package}
\title{tropo: Create Raster Tiles}
\description{
Create raster tiles abstractly, or from large input files.
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
